Move db operations into 'storage' module
